FCBHDBP-278 It has created the Plan services to share business logic between the controllers and the commands. It has the new services: PlanService and PlaylistService. The above will have the business logic to handle the translate action for the plan and playlist.
