+++ app/Console/Commands/TranslatePlan.php
@@ -7,13 +7,15 @@ use App\Models\Language\Language;
 use App\Models\Plan\Plan;
 use App\Models\Playlist\PlaylistItems;
 use App\Models\User\User;
+use App\Models\Bible\Bible;
+use App\Services\Plans\PlanService;
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
-class translatePlan extends Command
+class TranslatePlan extends Command
 {
     /**
      * The name and signature of the console command.
@@ -29,6 +31,8 @@ class translatePlan extends Command
      */
     protected $description = 'Command to translate a plan to a list of bibles';
 
+    private $plan_service;
+    
     /**
      * Create a new command instance.
      *
@@ -37,6 +41,7 @@ class translatePlan extends Command
     public function __construct()
     {
         parent::__construct();
+        $this->plan_service = new PlanService();
     }
 
     /**
@@ -49,10 +54,6 @@ class translatePlan extends Command
         $plan_id = $this->argument('plan_id');
         $bible_ids = $this->argument('bible_ids');
 
-
-
-        // i18n
-
         $cache_params = ['eng'];
         $current_language = cacheRemember('selected_api_language', $cache_params, now()->addDay(), function () {
             $language = Language::where('iso', 'eng')->select(['iso', 'id'])->first();
@@ -64,30 +65,27 @@ class translatePlan extends Command
         $GLOBALS['i18n_iso'] = $current_language['i18n_iso'];
         $GLOBALS['i18n_id']  = $current_language['i18n_id'];
 
+        $plan = $this->plan_service->getPlanById($plan_id);
 
-        $plan = Plan::where('id', $plan_id)->first();
-
-        $this->alert('Translating plan ' . $plan->name . ' starting: ' . Carbon::now());
         if (!$plan) {
             $this->error('Plan with ID:' . $plan_id . ' do not exist');
         } else {
+            $this->alert('Translating plan ' . $plan->name . ' starting: ' . Carbon::now());
             $bible_ids = explode(',', $bible_ids);
-            $user = User::with('projects')->whereId($plan->user_id)->first();
-            $plan_controller = new MessagedPlansController();
+
             foreach ($bible_ids as $key => $bible_id) {
-                $request = new Request();
-                request()->merge(['bible_id' => $bible_id]);
                 try {
                     $this->line('Translating plan to bible ' . $bible_id . ' started ' . Carbon::now());
-                    $new_plan = $plan_controller->translate($request, $plan_id, $user, false, false);
-                    $plan = Plan::where('id', $new_plan['id'])->first();
-                    $language_id = DB::connection('dbp')
-                        ->table('bibles')
-                        ->select(['language_id'])
-                        ->whereId($bible_id)->first();
-
-                    $plan->language_id = $language_id->language_id;
-                    $plan->save();
+                    $bible = cacheRemember(
+                        'bible_translate',
+                        [$bible_id],
+                        now()->addDay(),
+                        function () use ($bible_id) {
+                            return Bible::whereId($bible_id)->first();
+                        }
+                    );
+                    $translated_plan = $this->plan_service->translate($plan_id, $bible, $plan->user_id, false, false);
+                    $plan = Plan::where('id', $translated_plan['id'])->first();
 
                     $this->line('Calculating duration and verses ' . Carbon::now());
                     foreach ($plan->days as $day) {
@@ -99,6 +97,8 @@ class translatePlan extends Command
                     }
 
                     $this->info('Translating plan to bible ' . $bible_id . ' finalized ' . Carbon::now());
+                    $this->info('Plan Translated ID: ' . $plan->id . ' ' . Carbon::now());
+                    $this->info('Plan Translated Language ID: ' . $plan->language_id . ' ' . Carbon::now());
                     $this->line('');
                 } catch (Exception $e) {
                     $this->error('Error translating plan to bible ' . $bible_id . ' ');
@@ -109,7 +109,12 @@ class translatePlan extends Command
                     $this->question('To continue the process please run the following command: ');
                     $this->line('');
 
-                    $this->comment("\t<fg=green>php artisan translate:plan " . $plan_id . ' ' . implode(',', array_splice($bible_ids, $key)));
+                    $this->comment(
+                        "\t<fg=green>php artisan translate:plan " . $plan_id . ' ' . implode(
+                            ',',
+                            array_splice($bible_ids, $key)
+                        )
+                    );
                     break;
                 }
             }
@@ -117,14 +122,5 @@ class translatePlan extends Command
 
         $this->line('');
         $this->alert('Translating plan end: ' . Carbon::now());
-    }
-}
-
-
-class MessagedPlansController extends PlansController
-{
-    public function replyWithError($message, $action = null)
-    {
-        throw (new Exception($message));
     }
 }
