+++ app/Transformers/PlanDayPlaylistItemsTransformer.php
@@ -23,12 +23,17 @@ class PlanDayPlaylistItemsTransformer extends PlanTransformerBase
             "start_date" => $plan->start_date,
             "percentage_completed" => $plan->percentage_completed,
             "days" => $plan->days->map(function ($day) {
-                return [
+                $day_result = [
                     "id" => $day->id,
                     "playlist_id" => $day->playlist_id,
-                    "playlist" => $this->parsePlaylistData($day->playlist),
                     "completed" => $day->completed
                 ];
+
+                if ($this->params['show_details']) {
+                    $day_result["playlist"] = $this->parsePlaylistData($day->playlist);
+                }
+
+                return $day_result;
             }),
             "user" => [
                 "id" => $plan->user->id,
